+++ fwb/components/ui/privategroups/BargainsPicture.tsx
